Separate search input component from Header
