+++ client/src/components/layout/Header.jsx
@@ -1,62 +1,18 @@
 import { useState } from 'react';
 
 // @Material UI
-import { styled, alpha } from '@mui/material/styles';
 import AppBar from '@mui/material/AppBar';
 import Drawer from '@mui/material/Drawer';
 import Box from '@mui/material/Box';
 import Toolbar from '@mui/material/Toolbar';
 import IconButton from '@mui/material/IconButton';
 import Typography from '@mui/material/Typography';
-import InputBase from '@mui/material/InputBase';
 import MenuIcon from '@mui/icons-material/Menu';
-import SearchIcon from '@mui/icons-material/Search';
 import Link from '@mui/material/Link';
+import SearchInput from '../UI/SearchInput';
 
 // import assets
 import logo from '/images/delsuites-logo.svg';
-
-const Search = styled('div')(({ theme }) => ({
-  position: 'relative',
-  borderRadius: theme.shape.borderRadius,
-  backgroundColor: alpha(theme.palette.common.white, 0.15),
-  '&:hover': {
-    backgroundColor: alpha(theme.palette.common.white, 0.25),
-  },
-  marginLeft: 'auto',
-  width: '100%',
-  [theme.breakpoints.up('sm')]: {
-    marginLeft: 'auto',
-    width: 'auto',
-  },
-}));
-
-const SearchIconWrapper = styled('div')(({ theme }) => ({
-  padding: theme.spacing(0, 2),
-  height: '100%',
-  position: 'absolute',
-  pointerEvents: 'none',
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-}));
-
-const StyledInputBase = styled(InputBase)(({ theme }) => ({
-  color: 'inherit',
-  '& .MuiInputBase-input': {
-    padding: theme.spacing(1, 1, 1, 0),
-    // vertical padding + font size from searchIcon
-    paddingLeft: `calc(1em + ${theme.spacing(4)})`,
-    transition: theme.transitions.create('width'),
-    width: '100%',
-    [theme.breakpoints.up('sm')]: {
-      width: '12ch',
-      '&:focus': {
-        width: '20ch',
-      },
-    },
-  },
-}));
 
 const Header = () => {
   const [menuActive, setMenuActive] = useState(false);
@@ -102,15 +58,7 @@ const Header = () => {
             delsuites
           </Typography>
         </Link>
-        <Search sx={{ display: { xs: 'none', md: 'block' } }}>
-          <SearchIconWrapper>
-            <SearchIcon />
-          </SearchIconWrapper>
-          <StyledInputBase
-            placeholder='Search…'
-            inputProps={{ 'aria-label': 'search' }}
-          />
-        </Search>
+        <SearchInput sx={{ display: { xs: 'none', md: 'block' } }} />
       </Toolbar>
       <Drawer
         open={menuActive}
@@ -119,15 +67,7 @@ const Header = () => {
         }}
       >
         <Box sx={{ p: 3 }}>
-          <Search sx={{ display: { xs: 'block', md: 'none' } }}>
-            <SearchIconWrapper>
-              <SearchIcon />
-            </SearchIconWrapper>
-            <StyledInputBase
-              placeholder='Search…'
-              inputProps={{ 'aria-label': 'search' }}
-            />
-          </Search>
+          <SearchInput sx={{ display: { xs: 'block', md: 'none' } }} />
           <ul>
             <li>
               <Typography>Menu Item</Typography>
